fixing capabilities checks

+++ classes/external.php
@@ -206,8 +206,10 @@ class external extends \external_api {
         // Make sure the user has the required capabilities.
         $context = \context_system::instance();
         self::validate_context($context);
-        // TODO wrong capability check, use API.
-        require_capability('tool/certificate:manage', $context);
+        $template = \tool_certificate\template::find_by_id($issue->templateid);
+        if (!$template->can_issue()) {
+            throw new \required_capability_exception($template->get_context(), 'tool/certificate:issue', 'nopermissions', 'error');
+        }
 
         // Delete the issue.
         return $DB->delete_records('tool_certificate_issues', ['id' => $issue->id]);
